Module nodle_staking::chain_spec

Structs

Extensions

Node ChainSpec extensions.

GenesisConfig

Functions

authority_keys_from_seed

Helper function to generate stash, controller and session key from seed

build_local_properties

development_config

Development config (single validator Alice)

development_config_genesis

get_account_id_from_seed

Helper function to generate an account ID from seed

get_from_seed

Helper function to generate a crypto pair from seed

local_staking_config

Local testnet config to test the staking pallet

local_staking_genesis

session_keys

staking_config

testnet_genesis

Helper function to create GenesisConfig for testing

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountPublic

Balance

Balance of an account.

Block

Block type.

BlockNumber

An index to a block.

ChainSpec

Specialized ChainSpec.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.